requirements.txt
streamlit==1.22.0
transformers==4.30.2
pillow==9.1.1